refactor: FigureFactory 추출
Figure의 인스턴스 생성 기능을 팩토리로 분리.
