+++ src/main/java/coord/model/FigureCreator.java
@@ -1,0 +1,5 @@
+package coord.model;
+
+public interface FigureCreator {
+    Figure create(Points points);
+}
