+++ src/test/java/coord/model/QuadrangleTest.java
@@ -13,7 +13,7 @@ class QuadrangleTest {
         Points points = new Points(Arrays.asList(new Point(1, 10),
                 new Point(1, 1), new Point(5, 1), new Point(5, 10)));
         assertEquals(36.0,
-                Figure.getInstance(points).area(), 0.001);
+                FigureFactory.getInstance(points).area(), 0.001);
     }
 
     @Test
@@ -21,6 +21,6 @@ class QuadrangleTest {
         Points points = new Points(Arrays.asList(new Point(1, 1),
                 new Point(10, 7), new Point(15, 10), new Point(3, 20)));
 
-        assertEquals(122.5, Figure.getInstance(points).area(), 0.1);
+        assertEquals(122.5, FigureFactory.getInstance(points).area(), 0.1);
     }
 }
